More Data Type Correction for `HelpData`
Fixed the data types for the maps in `HelpData` to be Long instead of
Integer, as the Stream operations return longs.

+++ app/src/main/java/xterminators/spellingbee/model/HelpData.java
@@ -25,6 +25,6 @@ public record HelpData(
     int totalPoints,
     long numPangrams,
     long numPerfectPangrams,
-    Map<Pair<Character, Integer>, Integer> startingLetterGrid,
-    Map<String, Integer> startingLetterPairs
+    Map<Pair<Character, Integer>, Long> startingLetterGrid,
+    Map<String, Long> startingLetterPairs
 ) {}
